+++ src/components/DetailHeader.js
@@ -1,13 +1,16 @@
 import { useState, useLayoutEffect } from 'react';
 import { useParams } from 'react-router-dom';
 import { useSelector, useDispatch } from 'react-redux';
-
+import { BsHeart, BsHeartFill } from 'react-icons/bs';
 import {
   addUserFavoriteBuilding,
   deleteUserFavoriteBuilding,
 } from '../store/userSlice';
 
+import { Agent } from '../utils/constants';
+
 import useAxios from '../hooks/useAxios';
+import DropDown from './DropDown';
 
 export default function DetailHeader() {
   const dispatch = useDispatch();
@@ -62,6 +65,20 @@ export default function DetailHeader() {
           <div className='detail-auction-number'>
             {detail && <span>{auctionNumber}</span>}
           </div>
+          {userFavorite ? (
+            <button onClick={handleUserFavoriteRegion}>
+              <BsHeartFill></BsHeartFill>
+            </button>
+          ) : (
+            <button onClick={handleUserFavoriteRegion}>
+              <BsHeart></BsHeart>
+            </button>
+          )}
+          <DropDown
+            title={'권한위임'}
+            option={Agent}
+            auctionNumber={auctionNumber}
+          />
         </div>
       </div>
     </>
